Instantiation of db interface and git interface switched place
The database connection was established before interaction with git
started. Thus the database must exist before the git repository will be
cloned. But a dump of the database may exist in the git repository, so
one might want the git repository first, so the database can be restored
from the dump in the git repository. So it is better to pull or clone
the git repository first.